requirements.txt
fastapi[all]
uvicorn
sqlalchemy
alembic
pydantic
black
python-dotenv
fastapi-users[sqlalchemy2]
fastapi-users[oauth2]
asyncpg
pytest
pytest-asyncio
pydantic-settings
jinja2
aiofiles
aiocsv
babel
aiosqlite
aiosmtplib
authlib
pyjwt
slowapi
fastapi-users-db-sqlalchemy
httpx-oauth
